Move `EncodingSupport`
