+++ src/encode/mod.rs
@@ -1,6 +1,6 @@
-use std::io::Write;
+use std::{io::Write, num::NonZeroU8};
 
-use crate::{ColorFormat, EncodeError, Format, Size};
+use crate::{ColorFormat, EncodeError, Format, Size, SizeMultiple};
 
 mod bc;
 mod bc1;
@@ -233,4 +233,49 @@ impl Default for CompressionQuality {
     fn default() -> Self {
         Self::Normal
     }
+}
+
+/// Describes the extent of support for encoding a format.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct EncodingSupport {
+    /// Whether and what type of dithering is supported.
+    pub dithering: Dithering,
+    /// The split height for the image format.
+    ///
+    /// Encoding most formats is trivially parallelizable, by splitting the
+    /// image into chunks by lines, encoding each chunk separately, and writing
+    /// the encoded chunks to the output stream in order.
+    ///
+    /// This value specifies how many lines need to be grouped together for
+    /// correct encoding. E.g. `BC1_UNORM` requires 4 lines to be grouped
+    /// together, meaning that all chunks (except the last one) must have a
+    /// height that is a multiple of 4. So e.g. an image with a height of 10
+    /// pixels can split into chunks with heights of 4-4-2, 8-2, 4-6, or 10.
+    ///
+    /// Note that most dithering will produce different (but not necessarily
+    /// incorrect) results if the image is split into chunks. However, all BCn
+    /// formats implement block-based local dithering, meaning that the dithering
+    /// is the same whether the image is split or not. See
+    /// [`EncodingSupport::local_dithering`].
+    pub split_height: Option<NonZeroU8>,
+    /// Whether the format supports local dithering.
+    ///
+    /// Most formats implement global error diffusing dithering for best quality.
+    /// However, this prevents parallel encoding of the image, as the dithering
+    /// error of one chunk depends on the dithering error of the previous chunk.
+    /// It's still possible to encode the image in parallel, but the dither
+    /// pattern may reveal the chunk seams.
+    ///
+    /// Local dithering on the other hand will attempt to diffuse the error
+    /// within a small region of the image. E.g. `BC1_UNORM` will dither within
+    /// a 4x4 block. This allows the image to be split into chunks and encoded
+    /// in parallel without revealing the chunk seams.
+    ///
+    /// `self.dithering == Dithering::None` implies `self.local_dithering == false`.
+    pub local_dithering: bool,
+    /// The size multiple of the encoded image.
+    ///
+    /// If the dimensions of the image are not multiples of this size, the
+    /// encoder with return an error.
+    pub size_multiple: SizeMultiple,
 }
