+++ src/lib.rs
@@ -21,7 +21,9 @@ use std::num::NonZeroU8;
 pub use color::*;
 pub use decode::{decode, decode_rect, DecodeOptions};
 pub use decoder::*;
-pub use encode::{encode, CompressionQuality, Dithering, EncodeOptions, ErrorMetric};
+pub use encode::{
+    encode, CompressionQuality, Dithering, EncodeOptions, EncodingSupport, ErrorMetric,
+};
 pub use encoder::*;
 pub use error::*;
 pub use format::*;
